requirements.txt
docker-py==1.2.2
click==4.0
